+++ pywb/framework/test/test_wsgi_wrapper.py
@@ -22,7 +22,7 @@ class TestCustomErrApp:
 
 
 def initer(app_class):
-    def init():
+    def init(config=None):
         return app_class()
     return init