+++ pywb/webapp/pywb_init.py
@@ -6,13 +6,10 @@ from pywb.framework.wbrequestresponse import WbRequest
 from pywb.framework.memento import MementoRequest
 from pywb.framework.basehandlers import BaseHandler
 
-from pywb.warc.recordloader import ArcWarcRecordLoader
-from pywb.warc.resolvingloader import ResolvingLoader
-
-from views import J2TemplateView, add_env_globals
+from views import J2TemplateView
 from views import J2HtmlCapturesView, HeadInsertView
 
-from replay_views import ReplayView
+from live_rewrite_handler import RewriteHandler
 
 from query_handler import QueryHandler
 from handlers import WBHandler
@@ -62,31 +59,10 @@ class DictChain:
 
 #=================================================================
 def create_wb_handler(query_handler, config):
-
-    cookie_maker = config.get('cookie_maker')
-    record_loader = ArcWarcRecordLoader(cookie_maker=cookie_maker)
-
-    paths = config.get('archive_paths')
-
-    resolving_loader = ResolvingLoader(paths=paths,
-                                       record_loader=record_loader)
-
-    template_globals = config.get('template_globals')
-    if template_globals:
-        add_env_globals(template_globals)
-
-    replayer = ReplayView(resolving_loader, config)
-
-    search_view = (J2TemplateView.
-                   create_template(config.get('search_html'),
-                                   'Search Page'))
-
     wb_handler_class = config.get('wb_handler_class', WBHandler)
 
     wb_handler = wb_handler_class(
         query_handler,
-        replayer,
-        search_view=search_view,
         config=config,
     )
 
@@ -94,23 +70,33 @@ def create_wb_handler(query_handler, config):
 
 
 #=================================================================
-def init_collection(value, config):
+def create_live_handler(config):
+    live_handler = RewriteHandler(config)
+    return live_handler
+
+
+#=================================================================
+def init_route_config(value, config):
     if isinstance(value, str):
-        value = {'index_paths': value}
+        value = dict(index_paths=value)
 
     route_config = DictChain(value, config)
+    return route_config
 
+
+#=================================================================
+def init_collection(route_config):
     ds_rules_file = route_config.get('domain_specific_rules', None)
 
     html_view = (J2HtmlCapturesView.
-                 create_template(config.get('query_html'),
+                 create_template(route_config.get('query_html'),
                                  'Captures Page'))
 
     query_handler = QueryHandler.init_from_config(route_config,
                                                   ds_rules_file,
                                                   html_view)
 
-    return route_config, query_handler
+    return query_handler
 
 
 #=================================================================
@@ -139,8 +125,8 @@ def create_cdx_server_app(passed_config):
     routes = []
 
     for name, value in collections.iteritems():
-        result = init_collection(value, config)
-        route_config, query_handler = result
+        route_config = init_route_config(value, config)
+        query_handler = init_collection(route_config)
 
         cdx_api_suffix = route_config.get('enable_cdx_api', True)
 
@@ -173,22 +159,31 @@ def create_wb_router(passed_config={}):
     else:
         request_class = WbRequest
 
-    #if config.get('use_lxml_parser', False):
-    #    use_lxml_parser()
+    # store live and replay handlers
+    handler_dict = {}
 
     for name, value in collections.iteritems():
-
         if isinstance(value, BaseHandler):
+            handler_dict[name] = value
             routes.append(Route(name, value))
             continue
 
-        result = init_collection(value, config)
-        route_config, query_handler = result
+        route_config = init_route_config(value, config)
+
+        if route_config.get('index_paths') == '$liveweb':
+            live = create_live_handler(route_config)
+            handler_dict[name] = live
+            routes.append(Route(name, live))
+            continue
+
+        query_handler = init_collection(route_config)
 
         wb_handler = create_wb_handler(
             query_handler=query_handler,
-            config=route_config
+            config=route_config,
         )
+
+        handler_dict[name] = wb_handler
 
         logging.debug('Adding Collection: ' + name)
 
@@ -215,6 +210,12 @@ def create_wb_router(passed_config={}):
     for static_name, static_path in static_routes.iteritems():
         routes.append(Route(static_name, StaticHandler(static_path)))
 
+    # resolve any cross handler references
+    for route in routes:
+        if hasattr(route.handler, 'resolve_refs'):
+            route.handler.resolve_refs(handler_dict)
+
+
     # Check for new proxy mode!
     if config.get('enable_http_proxy', False):
         router = ProxyArchivalRouter
@@ -237,5 +238,7 @@ def create_wb_router(passed_config={}):
                                                  'Home Page'),
 
         error_view=J2TemplateView.create_template(config.get('error_html'),
-                                                 'Error Page')
+                                                 'Error Page'),
+
+        config=config
     )