+++ sample_archive/text_content/toptest.js
@@ -1,0 +1,1 @@
+!function(){top!==window&&(alert("For security reasons, framing is not allowed."),top.location.replace(document.location))}